ArduDino's Python Script.
The main idea for this script is to recognize the 'keyword' sent by the Arduino/C++ script, and use PyAutoGUI to perform the appropriate function on the famous Dino Game. The IR remote sends a hexadecimal signal to the IR receiver, which becomes interpreted as either 'jump' or 'duck' to the Python script by COM3 port communication from the Arduino, activating the appropriate PyAutoGUI functions to perform the 'jump' action on the Dino in the game.