ArduDino's Arduino Script
The main idea for this script is to provide functionality to the IR remote and receiver that is wired up to the breadboard. The IR remote sends a signal to 'jump' or 'duck' to the Python script, which performs the respective action on the dinosaur in the popular 'Dino Game'.